requirements.txt
pygame>=2.5.0
numpy>=1.24.0
requests>=2.31.0
Pillow>=10.0.0
